lift chart state up and implement chart toggle
