+++ src/Components/Dashboard/Dashboard.js
@@ -2,10 +2,7 @@ import React from 'react';
 import Doughnut from '../charts/Doughnut';
 import Transaction from './PastTransaction';
 import Wallet from './Wallet';
-import PerformanceLine from '../charts/PerformanceLine';
-import PerformanceBar from '../charts/PerformanceBar';
-import PerformanceMultiLine from '../charts/PerformanceMultiLine';
-import LineChart from '../LineChart'
+import Charts from './Charts';
 import Currency from './CurrentUserCurrency';
 
 import './dashboard.scss';
@@ -15,7 +12,7 @@ const Dashboard = () => {
     <div className="dashboard-container">
       <section className="top-section">
         {/* https://www.chartjs.org/docs/latest/general/responsive.html */}
-        <div className="chart-container">
+        <div className="doughnut-container">
           <Doughnut />
         </div>
 
@@ -25,12 +22,8 @@ const Dashboard = () => {
       </section>
 
       <section className="middle-section">
-        <div className="pc1">
-          <LineChart />
-        </div>
-
-        <div className="pc2">
-          <PerformanceBar />
+        <div className="chart-container">
+          <Charts />
         </div>
       </section>
