+++ src/Components/Dashboard/dashboard.scss
@@ -14,7 +14,7 @@
   justify-content: flex-start;
 }
 
-.chart-container {
+.doughnut-container {
   width: 40vw;
 }
 
@@ -26,16 +26,9 @@
   width: 60vw;
 }
 
-.pc1, .pc2 {
-  position: relative;
-  //using vh messes with chart on window-resize
-  width: 450px !important;
-  height: 300px !important;
-}
-
-.pc2 {
-  margin-top: 5em;
-  margin-left: 2em;
+// define the container properties here
+.chart-container {
+  width: 700px;
 }
 
 .transactions {
