+++ src/Components/market/TradingWidget.js
